code update

+++ projects/storefrontlib/src/cms-components/cart/cart-details/cart-details.component.ts
@@ -1,11 +1,19 @@
-import { ChangeDetectionStrategy, Component, OnInit } from '@angular/core';
+import {
+  ChangeDetectionStrategy,
+  Component,
+  OnInit,
+  OnDestroy,
+} from '@angular/core';
 import {
   Cart,
   CartService,
   OrderEntry,
   SelectiveCartService,
+  AuthService,
+  RoutingService,
+  AuthRedirectService,
 } from '@spartacus/core';
-import { Observable, combineLatest } from 'rxjs';
+import { Observable, combineLatest, Subscription } from 'rxjs';
 import { filter, map } from 'rxjs/operators';
 import { Item } from '../cart-shared/cart-item/cart-item.component';
 
@@ -14,14 +22,19 @@ import { Item } from '../cart-shared/cart-item/cart-item.component';
   templateUrl: './cart-details.component.html',
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class CartDetailsComponent implements OnInit {
+export class CartDetailsComponent implements OnInit, OnDestroy {
   cart$: Observable<Cart>;
   entries$: Observable<OrderEntry[]>;
   cartLoaded$: Observable<boolean>;
 
+  private subscription: Subscription;
+
   constructor(
     protected cartService: CartService,
-    protected selectiveCartService: SelectiveCartService
+    protected selectiveCartService: SelectiveCartService,
+    private authService: AuthService,
+    private routingService: RoutingService,
+    private authRedirectService: AuthRedirectService
   ) {}
 
   ngOnInit() {
@@ -48,7 +61,22 @@ export class CartDetailsComponent implements OnInit {
   }
 
   saveForLater(item: Item) {
-    this.cartService.removeEntry(item);
-    this.selectiveCartService.addEntry(item.product.code, item.quantity);
+    this.subscription = this.authService
+      .isUserLoggedIn()
+      .subscribe(loggedIn => {
+        if (loggedIn) {
+          this.cartService.removeEntry(item);
+          this.selectiveCartService.addEntry(item.product.code, item.quantity);
+        } else {
+          this.routingService.go({ cxRoute: 'login' });
+          this.authRedirectService.reportAuthGuard();
+        }
+      });
+  }
+
+  ngOnDestroy(): void {
+    if (this.subscription) {
+      this.subscription.unsubscribe();
+    }
   }
 }
