+++ projects/storefrontlib/src/cms-components/cart/save-for-later/save-for-later.component.ts
@@ -1,11 +1,14 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, OnDestroy } from '@angular/core';
 import {
   SelectiveCartService,
   Cart,
   OrderEntry,
   ActiveCartService,
+  AuthService,
+  RoutingService,
+  AuthRedirectService,
 } from '@spartacus/core';
-import { Observable, combineLatest } from 'rxjs';
+import { Observable, combineLatest, Subscription } from 'rxjs';
 import { filter, map } from 'rxjs/operators';
 import { Item } from '../cart-shared/cart-item/cart-item.component';
 
@@ -13,14 +16,19 @@ import { Item } from '../cart-shared/cart-item/cart-item.component';
   selector: 'cx-save-for-later',
   templateUrl: './save-for-later.component.html',
 })
-export class SaveForLaterComponent implements OnInit {
+export class SaveForLaterComponent implements OnInit, OnDestroy {
   cart$: Observable<Cart>;
   entries$: Observable<OrderEntry[]>;
   cartLoaded$: Observable<boolean>;
 
+  private subscription: Subscription;
+
   constructor(
     protected cartService: ActiveCartService,
-    protected selectiveCartService: SelectiveCartService
+    protected selectiveCartService: SelectiveCartService,
+    private authService: AuthService,
+    private routingService: RoutingService,
+    private authRedirectService: AuthRedirectService
   ) {}
 
   ngOnInit() {
@@ -47,7 +55,22 @@ export class SaveForLaterComponent implements OnInit {
   }
 
   moveToCart(item: Item) {
-    this.selectiveCartService.removeEntry(item);
-    this.cartService.addEntry(item.product.code, item.quantity);
+    this.subscription = this.authService
+      .isUserLoggedIn()
+      .subscribe(loggedIn => {
+        if (loggedIn) {
+          this.selectiveCartService.removeEntry(item);
+          this.cartService.addEntry(item.product.code, item.quantity);
+        } else {
+          this.routingService.go({ cxRoute: 'login' });
+          this.authRedirectService.reportAuthGuard();
+        }
+      });
+  }
+
+  ngOnDestroy(): void {
+    if (this.subscription) {
+      this.subscription.unsubscribe();
+    }
   }
 }
